fix seed.ts and tsconfig.ts
The new seed.ts inside the prisma folder introduced an error because the file is outside of the previously set src root in tsconfig. So, I have changed the rootDir to include the whole project.

The seed.ts is also changed to import prisma client from the output set in schema file because it otherwise does not recognise the generated files